requirements.txt
fastapi>=0.112,<0.116
uvicorn[standard]>=0.23,<0.32
pydantic>=2.0,<3
requests>=2.31,<3
speedtest-cli>=2.1.3,<3
python-dotenv>=1.0,<2


